+++ src/components/input-field/input-field.tsx
@@ -470,11 +470,19 @@ export class InputField {
         this.isModified = true;
     };
 
+    private hasHelperText = () => {
+        return this.helperText !== null && this.helperText !== undefined;
+    };
+
+    private hasHelperLine = () => {
+        return this.maxlength || this.hasHelperText();
+    };
+
     private renderHelperLine = () => {
         const text: string = this.value || '';
         const length = text.length;
 
-        if (!this.maxlength && !this.hasHelperText()) {
+        if (!this.hasHelperLine()) {
             return;
         }
 
@@ -497,10 +505,6 @@ export class InputField {
                 </span>
             );
         }
-    };
-
-    private hasHelperText = () => {
-        return this.helperText !== null && this.helperText !== undefined;
     };
 
     private renderSuffix = () => {